+++ pkg/k8sclient/fake/clients.go
@@ -1,8 +1,6 @@
 package fake
 
 import (
-	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
-	versionedfake "github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned/fake"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
@@ -18,17 +16,14 @@ import (
 	clientfake "sigs.k8s.io/controller-runtime/pkg/client/fake" //nolint:staticcheck // v0.6.4 has a deprecation on pkg/client/fake that was removed in later versions
 
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
-	"github.com/giantswarm/k8sclient/v5/pkg/k8scrdclient"
 )
 
 type Clients struct {
 	logger micrologger.Logger
 
-	crdClient  k8scrdclient.Interface
 	ctrlClient client.Client
 	dynClient  dynamic.Interface
 	extClient  apiextensionsclient.Interface
-	g8sClient  versioned.Interface
 	k8sClient  kubernetes.Interface
 	restClient rest.Interface
 	restConfig *rest.Config
@@ -51,19 +46,6 @@ func NewClients(config k8sclient.ClientsConfig, objects ...runtime.Object) (*Cli
 		extClient = apiextensionsclientfake.NewSimpleClientset()
 	}
 
-	var crdClient *k8scrdclient.CRDClient
-	{
-		c := k8scrdclient.Config{
-			K8sExtClient: extClient,
-			Logger:       config.Logger,
-		}
-
-		crdClient, err = k8scrdclient.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var ctrlClient client.Client
 	{
 		if config.SchemeBuilder != nil {
@@ -78,17 +60,15 @@ func NewClients(config k8sclient.ClientsConfig, objects ...runtime.Object) (*Cli
 			}
 		}
 
-		ctrlClient = clientfake.NewFakeClientWithScheme(scheme.Scheme, objects...)
+		ctrlClient = clientfake.NewClientBuilder().
+			WithScheme(scheme.Scheme).
+			WithRuntimeObjects(objects...).
+			Build()
 	}
 
 	var dynClient dynamic.Interface
 	{
 		dynClient = dynamicfake.NewSimpleDynamicClient(scheme.Scheme, objects...)
-	}
-
-	var g8sClient versioned.Interface
-	{
-		g8sClient = versionedfake.NewSimpleClientset()
 	}
 
 	var k8sClient kubernetes.Interface
@@ -99,20 +79,14 @@ func NewClients(config k8sclient.ClientsConfig, objects ...runtime.Object) (*Cli
 	c := &Clients{
 		logger: config.Logger,
 
-		crdClient:  crdClient,
 		ctrlClient: ctrlClient,
 		dynClient:  dynClient,
 		extClient:  extClient,
-		g8sClient:  g8sClient,
 		k8sClient:  k8sClient,
 		restConfig: restConfig,
 	}
 
 	return c, nil
-}
-
-func (c *Clients) CRDClient() k8scrdclient.Interface {
-	return c.crdClient
 }
 
 func (c *Clients) CtrlClient() client.Client {
@@ -125,10 +99,6 @@ func (c *Clients) DynClient() dynamic.Interface {
 
 func (c *Clients) ExtClient() apiextensionsclient.Interface {
 	return c.extClient
-}
-
-func (c *Clients) G8sClient() versioned.Interface {
-	return c.g8sClient
 }
 
 func (c *Clients) K8sClient() kubernetes.Interface {